+++ packages/zosuss/src/SshSession.ts
@@ -12,7 +12,6 @@
 import {
     ICommandArguments,
     ICommandOptionDefinition,
-    IProfile,
     Logger
 } from "@zowe/imperative";
 import { ISshSession } from "./doc/ISshSession";
@@ -147,7 +146,7 @@ export class SshSession {
 
     /**
      * Given command line arguments, create an SSH session configuration object.
-     * @param {IProfile} args - The arguments specified by the user
+     * @param {ICommandArguments} args - The arguments specified by the user
      * @returns {ISshSession} - A session configuration to be used for an SSH session.
      */
     public static createSshSessCfgFromArgs(args: ICommandArguments): ISshSession {
@@ -156,46 +155,6 @@ export class SshSession {
             keyPassphrase: args.keyPassphrase,
             handshakeTimeout: args.handshakeTimeout
         };
-    }
-
-    /**
-     * Given a z/OS SSH profile, create a SSH Client Session.
-     * @static
-     * @deprecated Use SshSession.createSshSessCfgFromArgs & others
-     * @param {IProfile} profile - The SSH profile contents
-     * @returns {Session} - A session for usage in the SSH Client
-     */
-    public static createBasicSshSession(profile: IProfile): SshSession {
-        this.log.debug("Creating a z/OS SSH session from the profile named %s", profile.name);
-        return new SshSession({
-            hostname: profile.host,
-            port: profile.port,
-            user: profile.user,
-            password: profile.password,
-            privateKey: profile.privateKey,
-            keyPassphrase: profile.keyPassphrase,
-            handshakeTimeout: profile.handshakeTimeout
-        });
-    }
-
-    /**
-     * Given command line arguments, create a SSH Client Session.
-     * @static
-     * @deprecated Use SshSession.createSshSessCfgFromArgs & others
-     * @param {IProfile} args - The arguments specified by the user
-     * @returns {SshSession} - A session for usage in the SSH Client
-     */
-    public static createBasicSshSessionFromArguments(args: ICommandArguments): SshSession {
-        this.log.debug("Creating a z/OS SSH session from arguments");
-        return new SshSession({
-            hostname: args.host,
-            port: args.port,
-            user: args.user,
-            password: args.password,
-            privateKey: args.privateKey,
-            keyPassphrase: args.keyPassphrase,
-            handshakeTimeout: args.handshakeTimeout
-        });
     }
 
     /**